From:    Arjan.deVet@adv.IAEhv.nl (Arjan de Vet)
I've made a patch to read src and dst ACL's from a file (one entry per
line), the new commands are 'acl <name> src_file <file>' and 'acl <name>
dst_file <file>'.

Furthermore I added a small patch to implement the ACL lookup speedup
described above: a matched record in an ACL will be moved to the second
position of the list (the first position is somewhat more difficult). This
speedup does not work for negated ACL's (!<acl-name>) as the whole list
still needs to be searched. But for my purpose (a long list of IP addresses
authorized to use the proxy) it should be an improvement.