From:    Robert Side <rside@aiinc.bc.ca> Date:    Mon, 18 Jan 1999 17:48:00 GMT
To get squid version 2.1.Patch1 to compile and run on SCO Unix Version
3.2v4.2 you need to use this configure line (from a csh like shell)

env LIBS=-lintl ./configure --disable-poll

The env LIBS=-lintl is required so squid can find strftime. Should
configure check for the existence of strftime in the standard
libaries and if it can't find it, check libintl.a on SCO systems?