From Robert Collins:
The bug is really more of an unexpected feature: when testing for
content-type on reply mime headers, messages with no entitie will
never match a mime type, so restrictive access configurations will
not let them though. The bad thing is that 304 responses (along
with 1xx & 204) will never have a content-type header. Therefore
they never match and IMS responses to the client always fail. And
at the moment there is no acl for the status line, so we cannot
let the user decide...