Fix the persistent connection case.
Persistent connections are a bit special and I mistreated them in my
original commit.

In a nutshell, i treated the second  "client" (second request) as having
had no data read from it. In the persistent connection this isn't the case -
at least one read is performed (one read was all I saw, but..) and then
the connection is stalled whilst data pours in from the first connection.

Now I pay attention to whats in http->reqbuf for the second connection
and feed that to clientSendMoreData. Seems to work, too.



Adrian