Wccp2 config cleanup by Steven Wilton
Here's a small patch to wccp2 in squid 2.6.  It fixes a difference between
the code and the documentation in the default squid conf (proto= compared to
protocol=).  It also removes the "ports_defined" config variable, enables
the flag when ports are actually defined and checks to see that this flag is
set before the config is accepted.