--enable-linux-netfilter documentation cleanup by Reuben Farrelly
This is mostly a backport from an old patch I submitted to squid-3 a couple of
years ago, it fixes references to Linux-2.4 as Netfilter is actually a 2.4 *or
newer* feature.

Also fixed up the TPROXY URL which while more or less correct, is probably more
correct and less likely to break if we point to the next level up which has both
a 2.4 and 2.6 directory (that seem identical anyway at this stage).  It also has
some extra Redhat kernels in there that are not visible if we send the user
straight to the /2.4 directory.