Fix for reply headers >4k introduced by my last commit.
Unfortunately I didn't take into account that the reply size could be
greater than 4k (ie, one copy.) So temporarily I'm introducing the
following:

* Don't bother trying to reuse the partial body data in the first
  copy in clientSendMoreHeaderData(), just handle the headers that
  we've been given
* Then, in clientCheckHeaderDone(), the check for body data uses
  state->body_buf and state->body_size, which are both 0.
* So it writes out the headers in one comm_write(), then schedules
  another copy through clientWriteComplete().

This is suboptimal - sending headers, waiting for comm_write() to complete,
then sending body data, but it'll only stay this long for as long as it
takes me to remove the reply status+headers from the store memory. Then
the initial read will return headers and some data if available (at store
offset 0 for the data!) and we can once again do it all in one write().