MFC: Fix Negotiate authentication again. got broken in 2.6.STABLE7 (fix for Bug #1792)
The core to this problem was the AUTHENTICATE_STATE_FINISHED and it's odd
and non-obvious use and requirements. This patch eleminates this state by
moving the "user is now authenticated" logics to where the helper reply
is processed like one may expect.

This split design was quite likely a leftover from the NTLM challenge cache
used in Squid-2.5.

Merged changes:
2007/02/27 22:39:50 hno +58 -74   Fix Negotiate authentication again. got broken in 2.6.STABLE7 (fix for Bug #1792)