Ability to cache authentication based on source IP
This patch adds two new options enabling caching of authentication on the source IP,
primarily targeted as a workaround to NTLM authentication, but may be used
to avoid reauthentication when restarting the browser or to allow access
from applications not supporting authentication for some time after
successful authentication with the browser.

Adds two new squid.conf directives. authenticate_ip_shortcircuit_ttl and _access
setting the TTL for the auth cache and filtering when the auth cache may be used
respectively.