Import cleaned up version of the zph patch
This is a cleaned up version of the ZPH patch by Marin Stavrev, Evgeni Gechev
and Venkatesh K. Reddy. It combines all three rolles versions of the patch,
supporting both
 - IP ToS / Diffserv
 - Socket priority
 - IP Option
but exluding the ability to mirror incoming TOS on cache misses as it's
still a bit unclear how that should be done and support also missing in
standard kernels. (the way it's done in the original patch looks a little odd)

Original patches used as input:
http://zph.bratcheda.org/squid-2.6.STABLE2-ToS_Hit_ToS_Preserve.patch
http://zph.bratcheda.org/squid-2.5.STABLE3-hit_prio.diff
http://zph.bratcheda.org/squid-2.5.STABLE7-option-marking-zph.diff