Improve squid 2's handling of unknown HTTP methods.
- Replace the method_t enum with a method_t structure consisting of a
  method_code_t enum value and a method string.
- Remove extension method handling.
- Replace method-based acls with simple wordlists.

The method_code_t enum is pretty much the same as the old method_t enum with
the exception that all the extension method entries are removed and replaced
with a METHOD_OTHER code.  This allows us to have a set of pre-allocated
method_t structures for all known methods.  Unknown methods are allocated a new
method_t structure and the method string is stored in this.  One method_t
structure is created per request involving an unknown method and it is freed
along with the request structure.