Author: Mark Nottingham <mnot@pobox.com> Bug #2643: monitor-direct cache_peer option
cache_cf can parse a full URL in a cache_peer's monitorurl, but it always sends
the request to the peer being monitored. In some cases, it's useful to monitor
a URL on another host (e.g., when a more capable monitoring system is in place,
or when centralised control is desirable).

This patch adds a new cache_peer option, monitor-direct, which sends the
request directly to the host in the URL.