Collate and HTMLize the current Squid manual pages
This is expected to create a *-manuals.tar.gz bundle like the cfgman
bundle before it with raw and HTML versions of all man files currently
in the Squid sources.  For the purpose of providing web page versions
as authorative reference documentation.

TODO: Add man .8 pages for many of the helpers.