Author: Mark Nottingham <mnot@pobox.com> MFC: Bug #2647: Reprioritise override-* and stale-while-revalidate
In the "At this point the response is stale" section of refreshCheck(), Squid
does the stale-while-revalidate and stale-if-error checks before the checks for
override-expire and override-lastmod.

In some circumstances, this can cause undesireable side effects. For example, a
response with an invalid Expires header, an Expires header in the past, or
Cache-Control: max-age=0, in combination with

refresh_pattern . 1 100% 1 override-expire stale-while-revalidate=15

will, once the one minute window passes, go into a state where requests are
always served as TCP_STALE_HIT (with a corresponding TCP_ASYNC_MISS), as long
as a request comes in the 15 second stale-while-revalidate window and at least
once every 15 seconds thereafter.

The fix is to place the override-* checks before the stale-while-revalidate
logic, so that an already-stale response has the overridden freshness applied
before the stale checks happen.